surface: move impl to types/wlr_{sub,}compositor.c
