+++ types/meson.build
@@ -59,7 +59,6 @@ wlr_files += files(
 	'wlr_screencopy_v1.c',
 	'wlr_server_decoration.c',
 	'wlr_subcompositor.c',
-	'wlr_surface.c',
 	'wlr_switch.c',
 	'wlr_tablet_pad.c',
 	'wlr_tablet_tool.c',
